requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
aiofiles>=23.0.0             # Required by FastAPI StaticFiles
pydantic>=2.7.0
python-dotenv>=1.0.1

# ── LLM (Ollama) ──────────────────────────────────────────────────────────────
requests>=2.31.0             # HTTP client for legacy call_llm() helper
langchain-ollama>=0.1.0      # LangChain-compatible ChatOllama wrapper
langchain-core>=0.2.0        # BaseChatModel, HumanMessage, ToolMessage, @tool

# ── LangGraph ─────────────────────────────────────────────────────────────────
langgraph>=0.1.0

# ── HTTP client (for search tools) ────────────────────────────────────────────
httpx>=0.27.0                # Used by search.py and auth/routes.py

# ── ORM ───────────────────────────────────────────────────────────────────────
sqlmodel>=0.0.16             # SQLModel ORM — SQLAlchemy 2 + Pydantic, by FastAPI author


# ── ORM ───────────────────────────────────────────────────────────────────────
sqlmodel>=0.0.16             # SQLModel ORM — SQLAlchemy 2 + Pydantic, by FastAPI author


# ── PII Anonymization ────────────────────────────────────────────────────────
presidio-analyzer>=2.2.0     # Microsoft Presidio — NLP-powered PII detection
presidio-anonymizer>=2.2.0   # Microsoft Presidio — PII replacement engine
spacy>=3.7.0                 # NLP backend required by Presidio analyzer


# ── Auth ───────────────────────────────────────────────────────────────────────
authlib>=1.3.0               # OAuth 2.0 client (GitHub flow)
PyJWT>=2.8.0                 # JWT sign / verify (HS256)
itsdangerous>=2.1.0          # Starlette SessionMiddleware signing

